requirements.txt
amqp==5.1.0
arabic-reshaper==2.1.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.3
django-phonenumber-field==6.1.0
django-redis==5.2.0
django-split-settings==1.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
future==0.18.2
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.0
packaging==21.3
phonenumbers==8.12.44
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
redis==4.1.4
reportlab==3.6.6
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.5
